Moved XML operations to separate class
